Deleted deprecated OntologyClassExpression, OntologyClassNegatedExpression and Profile classes. Moved core.packet.Packet up to package top-level. Renamed Packet as PhenoPacked to make intention of library clear.
