--- src/main/java/org/monarchinitiative/ppk/model/ontology/OntologyNegatedClassExpression.java
@@ -1,8 +1,0 @@
-package org.monarchinitiative.ppk.model.ontology;
-
-@Deprecated
-public class OntologyNegatedClassExpression extends OntologyClassExpression {
-	
-	OntologyClassExpression complementOf;
-
-}
