--- src/main/java/org/monarchinitiative/ppk/model/packet/Profile.java
@@ -1,10 +1,0 @@
-package org.monarchinitiative.ppk.model.packet;
-
-import java.util.List;
-
-import org.monarchinitiative.ppk.model.meta.Association;
-
-@Deprecated
-public class Profile<T extends Association> {
-	List<T> profileMembers;
-}
