--- src/main/java/org/monarchinitiative/ppk/model/packet/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author cjm
- *
- */
-package org.monarchinitiative.ppk.model.packet;
