+++ src/main/java/org/monarchinitiative/ppk/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * @author cjm
+ *
+ */
+package org.monarchinitiative.ppk;
